Factor out context.ts.
